+++ R/measures_compute.R
@@ -584,9 +584,7 @@ compute_measures <- function(dgm_name, method, method_setting, measures = NULL, 
 
   # Define all available measures if not specified
   if (is.null(measures))
-    measures <- c("bias", "relative_bias", "mse", "rmse", "empirical_variance",
-                  "empirical_se", "coverage", "power", "mean_ci_width", "interval_score", "convergence",
-                  "positive_likelihood_ratio", "negative_likelihood_ratio")
+    measures <- measures()
 
   # Define measure functions
   measure_functions <- list(